Message passing in simple test works.  Read whole comment for issues.
I'm unsure if doNothing is the appropriate default message handler,
based on the way change_world chains together its continuations.

The main addition that was needed was the field getActivationRecord on
the return of bigBang, which keeps information about the world's
record on the object that flows into Racket code.  Then when the
world's value is passed back in as part of a mail, we can find the
activation record and look up the message listener from there.

The calls to msgListener may need to be deferred somehow until after
the current invocation of change_world is called, but I think the only
thing that will happen is that worlds will be redrawn in an unexpected
order.

The main thing to do now is write more tests to make sure this
implementation isn't crazy.