Basic support for communicating back to the server with server-msgs.
Worlds can declare a name --- they need it to send messages to the
server.

A few well-known urls are used to serve these requests:

/get/worldname
/store/worldname
/effect/worldname

And information that the server-side of the widget uses is put into
the POSTDATA of the request.  I'm flagrantly abusing Http verbs here
(everything is a POST).  Oh well.

Haven't tested getting responses back.  Next todo is to get the glue
going so that you can instantiate these things and tie them together
on the server.